+++ RELEASE_NOTES.md
@@ -1,6 +1,8 @@
 Release Notes
 =============
 ## 1.3.0-beta2
+* Container Instance: Support for secure parameters for environment variables
+* Deployment Scripts: Support for secure parameters for environment variables
 * Web App: Unmanaged Server Farm uses Resource Id for fully-qualified path.
 
 ## 1.3.0-beta1